ENH: Add "New model" action to default application menu.
